# docs/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "SaveStream API",
    "version": "0.1.0",
    "description": "Phase 0.5 frozen public contract for SaveStream Web/Mobile. Breaking changes require an explicit contract version change and coordinated frontend/backend update."
  },
  "servers": [
    {
      "url": "/v1",
      "description": "Version 1 base path"
    }
  ],
...
  "paths": {
...
    "/recordings": {
      "post": {
        "tags": [
          "Recordings"
        ],
        "summary": "Create recording",
        "operationId": "createRecording",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecordingRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recording"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "402": {
            "description": "Insufficient credits",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Recordings"
        ],
        "summary": "List recordings",
        "operationId": "listRecordings",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RecordingStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingList"
                }
              }
            }
          }
        }
      }
    },
    "/recordings/{recording_id}": {
      "get": {
        "tags": [
          "Recordings"
        ],
        "summary": "Get recording",
        "operationId": "getRecording",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/OpaqueId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recording"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Recordings"
        ],
        "summary": "Soft-delete recording",
        "operationId": "deleteRecording",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/OpaqueId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Delete accepted"
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Not deletable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/security/mfa": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get admin MFA status",
        "operationId": "getAdminMfaStatus",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMfaStatus"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/admin/security/mfa/setup": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Begin admin TOTP setup",
        "operationId": "beginAdminMfaSetup",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Setup details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMfaSetup"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/admin/security/mfa/enable": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Enable admin TOTP",
        "operationId": "enableAdminMfa",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminMfaCodeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Enabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMfaStatus"
                }
              }
            }
          },
          "401": {
            "description": "Invalid MFA code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Setup required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/admin/security/mfa/verify": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Verify admin TOTP or recovery code",
        "operationId": "verifyAdminMfa",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminMfaCodeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verified",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMfaStatus"
                }
              }
            }
          },
          "401": {
            "description": "Invalid MFA code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "MFA setup required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Client-generated UUID. Same key + different body returns 409 IDEMPOTENCY_KEY_REUSED.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
...
    },
    "schemas": {
      "OpaqueId": {
        "type": "string",
        "description": "Opaque server-generated identifier. Clients MUST NOT parse it; UUID or ULID may be used internally."
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "additionalProperties": false
      },
...
      "RecordingStatus": {
        "type": "string",
        "enum": [
          "queued",
          "resolving",
          "waiting_live",
          "recording",
          "processing",
          "uploading",
          "completed",
          "failed",
          "stop_requested",
          "stopped",
          "waiting_for_cloud_slot",
          "missed_no_cloud_slot"
        ]
      },
...
      "CreateRecordingRequest": {
        "type": "object",
        "required": [
          "source"
        ],
        "properties": {
          "source": {
            "$ref": "#/components/schemas/Source"
          },
          "max_duration_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 1
          },
          "quality": {
            "type": "string",
            "enum": [
              "best"
            ],
            "default": "best"
          },
          "container": {
            "type": "string",
            "enum": [
              "mp4"
            ],
            "default": "mp4"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}